# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PulseRecover API",
    "description": "AI Payment Reliability & Revenue Recovery Engine. Probabilistic AI proposes, deterministic policy decides, payment infrastructure executes, verification proves. All money amounts are integer paise (INR).",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/incidents/{incident_id}/investigate": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Investigate",
        "description": "Run (or return) an AI investigation for an incident.\n\nRuns the diagnosis first when none exists, then the reasoner, persists the\nrun in ``agent_reports`` and audits it. With ``force_refresh`` a fresh\nreport is produced even if one already exists.",
        "operationId": "investigate_api_v1_incidents__incident_id__investigate_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/InvestigateRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/investigation": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Investigation",
        "description": "Latest completed investigation report for an incident.",
        "operationId": "get_investigation_api_v1_incidents__incident_id__investigation_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "List Audit",
        "operationId": "list_audit_api_v1_audit_get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "real_test",
                "research",
                "all"
              ],
              "type": "string",
              "default": "real_test",
              "title": "Environment"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/export/audit": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Audit",
        "description": "Export audit trail \u2014 append-only record of every decision and action.",
        "operationId": "export_audit_api_v1_export_audit_get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "real_test",
                "research"
              ],
              "type": "string",
              "default": "real_test",
              "title": "Environment"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "csv",
                "json"
              ],
              "type": "string",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric"
            }
          },
          {
            "name": "detected_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected From"
            }
          },
          {
            "name": "detected_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detected To"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "real_test",
                "research"
              ],
              "type": "string",
              "default": "real_test",
              "title": "Environment"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiInferenceView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": "inference"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "supporting_fact_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Fact Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "statement",
          "confidence"
        ],
        "title": "AiInferenceView"
      },
      "AlternativeHypothesisView": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "cause": {
            "type": "string",
            "title": "Cause"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "diagnosis"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "cause",
          "confidence"
        ],
        "title": "AlternativeHypothesisView"
      },
...
      "AuditListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AuditListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HypothesisView": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "supporting_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supporting Evidence"
          }
        },
        "type": "object",
        "required": [
          "title",
          "confidence"
        ],
        "title": "HypothesisView",
        "description": "Legacy flat hypothesis (kept for consumers of the scaffold contract);\nderived from ai_inferences."
      },
...
      "IncidentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "IncidentListResponse"
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "INVESTIGATING",
          "DIAGNOSED",
          "RECOVERING",
          "RESOLVED",
          "CLOSED",
          "FALSE_POSITIVE"
        ],
        "title": "IncidentStatus"
      },
...
      "InvestigateRequest": {
        "properties": {
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "default": false
          }
        },
        "type": "object",
        "title": "InvestigateRequest"
      },
      "InvestigateResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "report": {
            "$ref": "#/components/schemas/InvestigationReportView"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "incident_id",
          "status",
          "started_at",
          "report"
        ],
        "title": "InvestigateResponse"
      },
      "InvestigationReportView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "observed_facts": {
            "items": {
              "$ref": "#/components/schemas/ObservedFactView"
            },
            "type": "array",
            "title": "Observed Facts"
          },
          "ai_inferences": {
            "items": {
              "$ref": "#/components/schemas/AiInferenceView"
            },
            "type": "array",
            "title": "Ai Inferences"
          },
          "recommended_actions": {
            "items": {
              "$ref": "#/components/schemas/RecommendedActionView"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "recommended_next_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendedActionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommended_candidates": {
            "items": {
              "$ref": "#/components/schemas/RecommendedActionView"
            },
            "type": "array",
            "title": "Recommended Candidates"
          },
          "alternative_hypotheses": {
            "items": {
              "$ref": "#/components/schemas/AlternativeHypothesisView"
            },
            "type": "array",
            "title": "Alternative Hypotheses"
          },
          "revenue_implications": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevenueImplicationsView"
              },
              {
                "type": "null"
              }
            ]
          },
          "uncertainties": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainties"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          },
          "escalated": {
            "type": "boolean",
            "title": "Escalated",
            "default": false
          },
          "escalation_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Escalation Reasons"
          },
          "degraded": {
            "type": "boolean",
            "title": "Degraded",
            "default": false
          },
          "degraded_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded Reasons"
          },
          "stripped_claims": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Stripped Claims"
          },
          "tools_called": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools Called"
          },
          "reasoner": {
            "type": "string",
            "title": "Reasoner",
            "default": "heuristic"
          },
          "diagnosis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/HypothesisView"
            },
            "type": "array",
            "title": "Hypotheses"
          },
          "generated_by": {
            "type": "string",
            "title": "Generated By",
            "default": "heuristic"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "id",
          "incident_id",
          "summary",
          "created_at"
        ],
        "title": "InvestigationReportView"
      },
      "ObservedFactView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "statement",
          "tool"
        ],
        "title": "ObservedFactView"
      },
...
      "PolicyOutcomePreview": {
        "properties": {
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "rules_matched": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules Matched"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "PolicyOutcomePreview"
      },
...
      "RecommendedActionView": {
        "properties": {
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "amount_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "INR"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "expected_recovery_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Recovery Paise"
          },
          "policy_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyOutcomePreview"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "action_type",
          "rationale"
        ],
        "title": "RecommendedActionView"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}